feat: refactor sync mode handling by moving mode-related functions to modes.go
