+++ event-forwarder/pkg/forwarder/sync_windowed.go
@@ -29,7 +29,11 @@ func (s *windowedStrategy) Run(ctx context.Context) error {
 
 	// Emit initial mode telemetry once
 	if !s.started {
-		f.emitTelemetry(telemetry.NewSyncModeChanged(f.currentSyncMode, "initial_mode"))
+		if f.tsink != nil {
+			f.tsink.EmitModeChanged(f.currentSyncMode, "initial_mode")
+		} else {
+			f.emitTelemetry(telemetry.NewSyncModeChanged(f.currentSyncMode, "initial_mode"))
+		}
 		s.started = true
 	}
